Converted to double and sound, other things too.

+++ bossmult.cs
@@ -11,10 +11,10 @@ namespace BossScale
      public class bossmult: GlobalNPC
     {
        
-        public static float bossmults = 1f;
+        public static double bossmults = 1.0;
         public override void SetDefaults(NPC npc)
         {
-            float max = (float)npc.lifeMax;
+            double max = (double)npc.lifeMax;
    
             if (npc.boss == true)
             {
